Reduce flickering when using video-below-window
Many changes, first, add property that says if we have a frame or
not. This can be used by the application to decide whether to clip or
not to "punch a hole" in the window. Don't reset the size when
pausing, and don't keep the last frame around when stopping. Some
other cleanups too.